Minor Game Start Changes
