+++ npcs/scienceoutpost/starboozenpc.npctype
@@ -3,7 +3,7 @@
   "baseType" : "fuoutpostcivilian",
   "damageTeam" : 1,
     "scriptConfig" : {
-	"behavior" : "lounger",
+	"behavior" : "fu_lounger",
         "offeredQuests" : [ "1boozekit", "2brew_a", "2brew_b", "2brew_c","2brew_d","3hops", "4malt", "5wort", "6beer", "7distill", "8spirits", "9merrymead" ],
         "turnInQuests" : [ "1boozekit", "3hops", "4malt", "5wort", "6beer", "7distill", "8spirits", "9merrymead" ],
