+++ objects/ancient/ancientconsole/fuancientconsole.object
@@ -68,8 +68,6 @@
   "uniqueId" : "ancientgate2",
   
   "requiresTricorder" : 1,
-  "tricorderCheckRange" : 10,
-  
   
   "scripts" : [ "/objects/ancient/ancientconsole/fuancientconsole.lua" ],
   "scriptDelta" : 20,
@@ -81,8 +79,5 @@
   "interactData" : "/interface/warping/gatewarp.config",
 
   "inactiveInteractAction" : "ScriptPane",
-  "inactiveInteractData" : "/interface/scripted/ancientconsole/ancientconsolegui.config",
-  
-  "noTricorderInteractAction" : "ShowPopup",
-  "noTricorderInteractData" : {"message" : "Without a ^orange;Personal Tricorder^reset; you cannot properly interface with this ancient console to activate it. Please complete the FU intro mission to obtain one."}
+  "inactiveInteractData" : "/interface/scripted/ancientconsole/ancientconsolegui.config"
 }
