+++ objects/minibiome/lush/poptopcavern.object
@@ -36,5 +36,9 @@
   "rooting" : true,
   
   "destination" : "InstanceWorld:cavedungeon:%s:%s",
-  "vanishTime" : 999999999999
+  "vanishTime" : 999999999999,
+  "tricorderCheckRange" : 10,
+  
+  "noTricorderInteractAction" : "ShowPopup",
+  "noTricorderInteractData" : {"message" : "Without a ^orange;Personal Tricorder^reset; you cannot properly interface with this ancient console to activate it. Please complete the FU intro mission to obtain one."}
 }
