+++ quests/fu_questlines/tutorial/vinj/create_matterassembler.questtemplate
@@ -12,7 +12,7 @@
     "questGiver" : {
       "type" : "entity",
       "indicator" : "/interface/quests/questreceiver.animation"
-    }  
+    }
   },
   "associatedMission" : "scienceoutpostfu",
   
@@ -26,6 +26,7 @@
       "makeWire" : "Use the ^green;Machining Table^reset; to create ^orange;Wire^reset;",
       "makeElectromagnet" : "Use the ^green;Machining Table^reset; to create an ^orange;Electromagnet^reset;",
       "cavern" : "Find the cavern aand investigate the ^orange;Energy Source^reset;"
-    }
+    },
+	"completionRange" : 500
   }
 }
